+++ doc/changelog.rst
@@ -5,6 +5,9 @@ Changelog
 **dev**
 
 * Element editor lets you change any elements settings, not just elements in loaded systems
+* Add option to do a rotation after sequencer has finished
+* Add option to toggle between parallel/perspective projective
+* Add option to write logging output to file
 
 **v140529**